community/identity-form-autofiller-python/src/requirements.txt
# https://pypi.org/project/google-cloud-documentai
google-cloud-documentai==1.5.0

# https://pypi.org/project/google-auth
google-auth==2.9.1

# https://pypi.org/project/Pillow
Pillow==9.2.0

# https://pypi.org/project/flask
Flask==2.2.1

# https://pypi.org/project/gunicorn
gunicorn==20.1.0
